Edit Views Commit

+++ app/src/main/java/com/example/businessv1/business/data/BusinessPagingSource.kt
@@ -4,6 +4,7 @@ import androidx.paging.PagingSource
 import com.example.businessv1.business.data.cache.CacheDataSource
 import com.example.businessv1.business.data.network.NetworkDataSource
 import com.example.businessv1.business.domain.model.Business
+import com.example.businessv1.business.domain.state.DataState
 import retrofit2.HttpException
 import java.io.IOException
 
@@ -22,18 +23,32 @@ constructor(
 
         return try {
             val response = networkDataSource.getList(query, position,BUSINESS_LIMIT)
-            if (position == 0){
-                cacheDataSource.deleteAll()
-                cacheDataSource.insertList(response.businesses)
+            var businessList:List<Business>
+            when(response){
+                is DataState.Success->{
+                    if (position == 0){
+                        cacheDataSource.deleteAll()
+                        cacheDataSource.insertList(response.data.businesses)
+                    }
+
+                    businessList = response.data.businesses
+
+                    LoadResult.Page(
+                        data = businessList,
+                        prevKey = if (position == BUSINESS_STARTING_PAGE_INDEX) null else position - BUSINESS_LIMIT,
+                        nextKey = if (businessList.isEmpty()) null else position + BUSINESS_LIMIT
+                    )
+                }
+                is DataState.Error->{
+                    LoadResult.Page(
+                        data = cacheDataSource.get(),
+                        prevKey = if (position == BUSINESS_STARTING_PAGE_INDEX) null else position - 1,
+                        nextKey = if (cacheDataSource.get().isEmpty()) null else position + BUSINESS_LIMIT
+                    )
+                }
+                DataState.Loading -> TODO()
             }
 
-            val businessList = response.businesses
-
-            LoadResult.Page(
-                data = businessList,
-                prevKey = if (position == BUSINESS_STARTING_PAGE_INDEX) null else position - BUSINESS_LIMIT,
-                nextKey = if (businessList.isEmpty()) null else position + BUSINESS_LIMIT
-            )
         } catch (exception: IOException){
           /* if (cacheDataSource.get().isEmpty()){
                LoadResult.Error(exception)
